Additional files for integration tests
